master-service/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
httpx>=0.27.0,<1.0.0
jinja2>=3.1.4
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-multipart>=0.0.9
sqlalchemy>=2.0.30,<3.0.0
psycopg[binary]>=3.2.0,<4.0.0
cloudinary>=1.44.0,<2.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
bcrypt>=4.0.1,<4.1.0
PyJWT>=2.9.0,<3.0.0
pypdf>=4.3.0,<5.0.0
python-docx>=1.1.2,<2.0.0
requests>=2.32.0,<3.0.0
